Fix warnings

+++ src/ngscopeclient/Session.cpp
@@ -43,11 +43,13 @@
 #include "MultimeterDialog.h"
 #include "PowerSupplyDialog.h"
 #include "RFGeneratorDialog.h"
-#include <fstream>
 
 #include "../scopehal/LeCroyOscilloscope.h"
 #include "../scopehal/MockOscilloscope.h"
 #include "../scopeprotocols/EyePattern.h"
+
+#include <fstream>
+#include <cinttypes>
 
 #ifdef _WIN32
 #include <windows.h>
@@ -544,7 +546,7 @@ bool Session::LoadWaveformDataForScope(
 		auto hist = m_history.GetHistory(time);
 		if(hist && (hist->m_history.find(scope) != hist->m_history.end()))
 		{
-			LogWarning("Session contains duplicate data for time %ld.%ld, discarding\n", time.first, time.second);
+			LogWarning("Session contains duplicate data for time %" PRId64 ".%" PRId64 ", discarding\n", static_cast<int64_t>(time.first), time.second);
 			continue;
 		}
